Multiplayer scoring system added (#19)
Score is calculated for each of the players, and displayed on each client's GUI

A negative (-) score means you are losing, and the value is the amount you need to gain to break even

+++ Assets/Scripts/Game/Behaviours/Managers/LevelSingleplayer.cs
@@ -23,7 +23,7 @@ namespace Game.Behaviours.Managers {
 
         private float highestPlatform;
 
-        private int platformIndex = 0;
+        private ushort platformIndex = 0;
 
         public static LevelSingleplayer Create(LevelSingleplayer prefab) {
             LevelSingleplayer level = Instantiate(prefab);
